feat: implement badge awarding logic, grading orchestration services, and session result query handling

+++ src/MathInsight.Modules.Grading_Analytics/Queries/GetSessionResult/GetSessionResultQueryHandler.cs
@@ -63,7 +63,7 @@ public sealed class GetSessionResultQueryHandler
                 if (!string.IsNullOrEmpty(a.Question.DifficultyId) &&
                     difficultyLevels.TryGetValue(a.Question.DifficultyId, out var level))
                 {
-                    difficultyLevel = level;
+                    difficultyLevel = (byte)level;
                 }
 
                 return new GradedAnswerDetailDto
